pyproject.toml
[project]
name = "discord-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.1",
    "apscheduler==3.11.0",
    "discord-py>=2.6.0",
    "google-api-python-client>=2.185.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "gspread-asyncio>=2.0.0",
    "pillow==11.3.0",
    "python-dotenv>=1.1.1",
]
[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff]
line-length = 88
